change user.yaml from OADC to G3DH
when we added the clopidogrel study to OADC (open access data commons), we changed the name to Gen3 Data Hub (since that study is not open access). We created a new user.yaml at the time. We are moving to use the g3dh user.yaml to prevent confusion in the commons-users repo